move mockNotifier() to test_utils
